requirements.txt
pyffish
tqdm
numpy
chess
python-shogi
